Move JS files to javascript folder. Delete React app in frontend folder. Config puma server to start on port 3000, and comment cors file.
