+++ app/javascript/reducers/greetingReducer.js
