+++ app/views/root/index.html.erb
@@ -1,1 +1,2 @@
+<h1>Root Index page</h1>
 <div id="root"></div>
