+++ package.json
@@ -11,7 +11,7 @@
   },
   "scripts": {
     "build": "webpack --config webpack.config.js",
-    "start": "webpack-dev-server --port 3001 --host 0.0.0.0"
+    "start": "react-scripts start"
   },
   "devDependencies": {
     "@babel/core": "^7.24.3",
