+++ webpack.config.js
@@ -16,13 +16,6 @@ module.exports = {
       },
     ],
   },
-  devServer: {
-    static: {
-      directory: path.join(__dirname, 'assets'),
-    },
-    host: '0.0.0.0',
-    port: 3001,
-  },
   output: {
     filename: '[name].js',
     sourceMapFilename: '[name].js.map',
